+++ config/setup.phpunit.php
@@ -1,0 +1,23 @@
+<?php
+
+/*
+ * This file is part of the Eventum (Issue Tracking System) package.
+ *
+ * @copyright (c) Eventum Team
+ * @license GNU General Public License, version 2 or later (GPL-2+)
+ *
+ * For the full copyright and license information,
+ * please see the COPYING and AUTHORS files
+ * that were distributed with this source code.
+ */
+
+return [
+    'database' => [
+        'hostname' => 'localhost',
+        'database' => 'eventum',
+        'username' => 'mysql',
+        'password' => '',
+        'port' => 3306,
+    ],
+    'admin_user' => 2,
+];