+++ src/Db/Adapter/YiiAdapter.php
@@ -63,7 +63,7 @@ class YiiAdapter extends PdoAdapterBase implements AdapterInterface
                     'dsn' => $this->getDsn($config),
                     'username' => $config['username'],
                     'password' => $config['password'],
-                    'charset' => $this->getCharset(),
+                    'charset' => isset($config['charset']) ? $config['charset'] : $this->getCharset(),
                 ],
             ],
         ];