+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## [3.5.0]
 
 Upgrading to 3.5.x versions requires that you upgrade to 3.2.0 version first.
+This version drops IRC Bot from Eventum Core, see #371
 
 - Fix corrupted note body on blocked multi-part emails (@balsdorf)
+- Drop IRC Bot from eventum core, available as eventum/irc-bot instead (@glensc, #371)
 
 [3.5.0]: https://github.com/eventum/eventum/compare/v3.4.2...master